requirements.txt
Flask==3.0.0
mongoengine==0.27.0
python-dotenv==1.0.0
pymongo==4.3.3
flask-cors
